+++ packages/terra-framework-docs/src/terra-dev-site/test/collapsible-menu-view/CollapsibleMenuViewSingleItem.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import CollapsibleMenuView from 'terra-collapsible-menu-view';
+
+const CollapsibleMenuViewSingleItem = () => (
+  <CollapsibleMenuView>
+    <CollapsibleMenuView.Item text="-----------------------------------------------------------------------------------------------------------------" key="button" />
+  </CollapsibleMenuView>
+);
+
+export default CollapsibleMenuViewSingleItem;