+++ packages/terra-framework-docs/src/terra-dev-site/test/date-time-picker/DateTimePickerDefault.test.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import DateTimePickerExampleTemplate from './common/DateTimePickerExampleTemplate';
+
+const DateTimePickerDefault = () => (
+  <DateTimePickerExampleTemplate />
+);
+
+export default DateTimePickerDefault;