+++ packages/terra-framework-docs/CHANGELOG.md
@@ -1,13 +1,16 @@
 # Changelog
 
-* Added
-  * Migrate docs from terra-menu
-
 ## Unreleased
+
+* Changed
+  * Migrated terra-collapsible-menu-view-docs
+  * Migrated docs from date-picker.
+  * Migrated docs from date-time-picker
 
 * Added
   * Migrated docs from terra-navigation-prompt.
   * Migrate docs from terra-navigation-layout
+  * Migrate docs from terra-menu
 
 ## 1.7.0 - (March 8, 2022)